% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{map_beta}
\alias{map_beta}
\alias{ins_ifx_cb}
\alias{hypo_cb}
\alias{hypo_epd}
\alias{hypo_cnt_cb}
\alias{dex_amount_callback}
\alias{ts_to_win_hours}
\alias{hirid_pharma_win}
\alias{mimic_presc_cort}
\alias{aumc_cortico}
\title{Callback functions for project-specific concept items}
\usage{
map_beta(beta)

ins_ifx_cb(ins, ...)

hypo_cb(glu, ...)

hypo_epd(hypo, min_dur = hours(6L), ...)

hypo_cnt_cb(hypo_epi, ...)

dex_amount_callback(dex, interval, ...)

ts_to_win_hours(dur)

hirid_pharma_win(dur)

mimic_presc_cort(x, val_var, dur_var, ...)

aumc_cortico(x, dur_var, ...)
}
\arguments{
\item{beta}{Vasopressor adjustement}

\item{ins}{Object containing insulin values}

\item{...}{Further arguments are ignored}

\item{glu}{Object containing glucose measurements}

\item{hypo}{Object containing a hypoglycemia indicator}

\item{min_dur}{Minimal episode duration interval}

\item{hypo_epi}{Object containing a hypoglycemia episode index}

\item{dex}{Object containing dextrose values}

\item{interval, dur}{Duration}

\item{x}{Data}

\item{val_var, dur_var}{Value and duration columns}
}
\description{
Callback functions for project-specific concept items
}
