% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sep.R
\name{sep3_info}
\alias{sep3_info}
\title{Sepsis utilities}
\usage{
sep3_info(source, pids = NULL, keep_components = TRUE, dat = NULL)
}
\arguments{
\item{source}{Data source name}

\item{pids}{Patient IDs}

\item{keep_components}{Flag indicating whether to return individual score
components}

\item{dat}{Data with SOFA score (computed if \code{NULL})}
}
\description{
Sepsis utilities
}
